refactor : 단건 스케줄 조회 타입 수정 Integer -> Long으로 DB에 맞게 수정

+++ scheduler/src/main/java/com/hyun/scheduler/controller/ScheduleController.java
@@ -40,7 +40,7 @@ public class ScheduleController {
   }
 
   @GetMapping("/schedule/{scheduleId}")
-  public ResponseEntity<ScheduleResponseDto> findScheduleById(@PathVariable Integer scheduleId) {
+  public ResponseEntity<ScheduleResponseDto> findScheduleById(@PathVariable Long scheduleId) {
     ScheduleResponseDto scheduleResponse = scheduleService.findScheduleById(scheduleId);
     return ResponseEntity.status(HttpStatus.OK).body(scheduleResponse);
   }
